INT: Move intention to annotation quick fix #489
