+++ src/main/resources/META-INF/plugin.xml
@@ -423,10 +423,6 @@
             <className>org.rust.ide.intentions.IfLetToMatchIntention</className>
             <category>Rust</category>
         </intentionAction>
-        <intentionAction>
-            <className>org.rust.ide.intentions.AddTurbofishIntention</className>
-            <category>Rust</category>
-        </intentionAction>
 
         <!-- Run Configurations -->
