--- src/main/resources/intentionDescriptions/AddTurbofishIntention/before.rs.template
@@ -1,3 +1,0 @@
-fn foo(a: i32, b: i32) {
-    let _ = str::parse<i32>("48")
-}
