--- src/main/resources/intentionDescriptions/AddTurbofishIntention/description.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-This intention adds turbofish for generic function call when is missed.
-</body>
-</html>
